+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changes to ltpa-scraper
 
+### 1.0.1 - (November 26, 2019)
+ - Bugfix: When verbose was active, the call to scrape() did never resolve
+ - Improvement: Use cookieJar to receive the cookies (from tough-cookie's dependency)
+
 ### 1.0.0 - (October 31, 2019)
  - Initial version